requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
prisma==0.11.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.1
psycopg2-binary==2.9.7